test: Update tests for extensions
